server/service/backend-service/backend-service-cubeia-wallet-adapter/src/main/java/com/cubeia/backend/firebase/CashGamesBackendAdapter.java
/**
 * Copyright (C) 2010 Cubeia Ltd <info@cubeia.com>
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

package com.cubeia.backend.firebase;

import static com.cubeia.backend.cashgame.dto.OpenSessionFailedResponse.ErrorCode.UNSPECIFIED_ERROR;

import java.math.BigDecimal;
import java.util.ArrayList;
import java.util.Collections;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.TreeMap;
import java.util.UUID;
import java.util.concurrent.atomic.AtomicLong;

import com.cubeia.backoffice.users.api.dto.User;
import com.cubeia.backoffice.users.client.UserServiceClient;
import com.cubeia.network.users.firebase.api.UserServiceContract;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import com.cubeia.backend.cashgame.CashGamesBackend;
import com.cubeia.backend.cashgame.PlayerSessionId;
import com.cubeia.backend.cashgame.TableId;
import com.cubeia.backend.cashgame.TransactionId;
import com.cubeia.backend.cashgame.dto.AllowJoinResponse;
import com.cubeia.backend.cashgame.dto.AnnounceTableRequest;
import com.cubeia.backend.cashgame.dto.AnnounceTableResponse;
import com.cubeia.backend.cashgame.dto.BalanceUpdate;
import com.cubeia.backend.cashgame.dto.BatchHandRequest;
import com.cubeia.backend.cashgame.dto.BatchHandResponse;
import com.cubeia.backend.cashgame.dto.CloseSessionRequest;
import com.cubeia.backend.cashgame.dto.HandResult;
import com.cubeia.backend.cashgame.dto.OpenSessionRequest;
import com.cubeia.backend.cashgame.dto.OpenSessionResponse;
import com.cubeia.backend.cashgame.dto.ReserveFailedResponse.ErrorCode;
import com.cubeia.backend.cashgame.dto.ReserveRequest;
import com.cubeia.backend.cashgame.dto.ReserveResponse;
import com.cubeia.backend.cashgame.dto.TransactionUpdate;
import com.cubeia.backend.cashgame.dto.TransferMoneyRequest;
import com.cubeia.backend.cashgame.exceptions.BatchHandFailedException;
import com.cubeia.backend.cashgame.exceptions.GetBalanceFailedException;
import com.cubeia.backend.cashgame.exceptions.OpenSessionFailedException;
import com.cubeia.backend.cashgame.exceptions.ReserveFailedException;
import com.cubeia.backoffice.accounting.api.NoSuchAccountException;
import com.cubeia.backoffice.accounting.api.UnbalancedTransactionException;
import com.cubeia.backoffice.wallet.api.config.AccountRole;
import com.cubeia.backoffice.wallet.api.dto.AccountBalanceResult;
import com.cubeia.backoffice.wallet.api.dto.report.TransactionRequest;
import com.cubeia.backoffice.wallet.api.dto.report.TransactionResult;
import com.cubeia.events.event.SystemEvent;
import com.cubeia.events.event.SystemEventType;
import com.cubeia.events.event.SystemLevels;
import com.cubeia.firebase.api.server.SystemException;
import com.cubeia.firebase.api.service.clientregistry.PublicClientRegistryService;
import com.cubeia.games.poker.common.money.Currency;
import com.cubeia.games.poker.common.money.Money;
import com.cubeia.network.wallet.firebase.api.WalletServiceContract;
import com.cubeia.network.wallet.firebase.domain.TransactionBuilder;
import com.cubeia.poker.domainevents.api.DomainEventsService;
import com.google.common.annotations.VisibleForTesting;
import com.google.common.collect.Maps;

/**
 * Adapter from the Backend Service Contract to the Cubeia Wallet Service.
 *
 * @author w
 */
public class CashGamesBackendAdapter implements CashGamesBackend {

    /**
     * Hardcoded licensee id, should be part of the open session request
     */
    public static final int LICENSEE_ID = 0;

    /**
     * Hardcoded game id, should be configurable or part of requests
     */
    public static final int GAME_ID = 1;

//    static final Long RAKE_ACCOUNT_USER_ID = -1000L;
//
//    static final Long PROMOTIONS_ACCOUNT_USER_ID = -2000L;

    private Logger log = LoggerFactory.getLogger(CashGamesBackendAdapter.class);

    private final AtomicLong idSequence = new AtomicLong(0);

    @VisibleForTesting
    protected WalletServiceContract walletService;

    protected AccountLookupUtil accountLookupUtil;
    
    protected PublicClientRegistryService clientRegistry;

    /**
     * Maps currency to rake account.
     */
    @VisibleForTesting
    protected Map<String, Long> systemRakeAccounts = Maps.newHashMap();

    /**
     * Maps currency to promotions account.
     */
    protected Map<String, Long> promotionsAccounts = Maps.newHashMap();

    protected DomainEventsService domainEventService;

    protected UserServiceContract userService;

    public CashGamesBackendAdapter(WalletServiceContract walletService, AccountLookupUtil accountLookupUtil,
                                   PublicClientRegistryService clientRegistry, DomainEventsService domainEventService,
                                   UserServiceContract userService) throws SystemException {
        this.walletService = walletService;
        this.accountLookupUtil = accountLookupUtil;
        this.clientRegistry = clientRegistry;
		this.domainEventService = domainEventService;
        this.userService = userService;
    }

    @Override
    public boolean isSystemShuttingDown() {
        return false;
    }

    private long nextId() {
        return idSequence.getAndIncrement();
    }

    @Override
    public String generateHandId() {
        return UUID.randomUUID().toString();
    }

    @Override
    public AllowJoinResponse allowJoinTable(int playerId) {
        log.warn("allow join not implemented, will always return ok");
        return new AllowJoinResponse(true, -1);
    }

    @Override
    public AnnounceTableResponse announceTable(AnnounceTableRequest request) {
        String uniqueId = UUID.randomUUID().toString();
        final AnnounceTableResponse response = new AnnounceTableResponse(new TableId(request.tableId, uniqueId));
        response.setProperty(CashGamesBackendService.MARKET_TABLE_REFERENCE_KEY, "CUBEIA-TABLE-ID::" + uniqueId);
        return response;
    }

    @Override
    public OpenSessionResponse openSession(final OpenSessionRequest request) throws OpenSessionFailedException {
        OpenSessionResponse response = null;
        try {
            Long walletSessionId = walletService.startSession(request.getOpeningBalance().getCurrencyCode(), LICENSEE_ID, request.getPlayerId(),
                    request.getObjectId(), GAME_ID, "unknown-" + request.getPlayerId(), request.getAccountName());

            PlayerSessionId sessionId = new PlayerSessionId(request.playerId, String.valueOf(walletSessionId));
            response = new OpenSessionResponse(sessionId, Collections.<String, String>emptyMap());
            log.debug("new session opened, oId = {}, pId = {}, sId = {}", new Object[]{request.getObjectId(), request.getPlayerId(), response.getSessionId()});

            if (request.openingBalance.getAmount().compareTo(BigDecimal.ZERO) > 0) {
                log.debug("Opening session requests a reservation with amount: " + request.openingBalance.getAmount());
                reserve(new ReserveRequest(response.getSessionId(), request.getOpeningBalance()));
            }

            return response;
        } catch (Exception e) {
            if (response != null) {
                log.error("Failed opening session", e);
                closeSession(new CloseSessionRequest(response.getSessionId()));
            }
            String msg = "error opening session for player " + request.getPlayerId() + ": " + e.getMessage();
            throw new OpenSessionFailedException(msg, e, UNSPECIFIED_ERROR);
        }
    }

    @Override
    public void closeSession(CloseSessionRequest request) {
        PlayerSessionId sid = request.getPlayerSessionId();
        long sessionAccountId = getWalletSessionIdByPlayerSessionId(sid);
        log.debug("Closing session account in wallet " + sessionAccountId);
        com.cubeia.backoffice.accounting.api.Money amountDeposited = walletService.endSessionAndDepositAll(LICENSEE_ID, sessionAccountId,
                "session ended by game " + GAME_ID + ", player id = " + sid.playerId);

        log.debug("wallet session {} closed for player {}, amount deposited: {}", new Object[]{sessionAccountId, sid.playerId, amountDeposited});
    }

    private long getWalletSessionIdByPlayerSessionId(PlayerSessionId sid) {
        return Long.valueOf(sid.integrationSessionId);
    }

    @Override
    public ReserveResponse reserve(final ReserveRequest request) throws ReserveFailedException {
        Money amount = request.getAmount();
        PlayerSessionId sid = request.getPlayerSessionId();
        Long walletSessionId = getWalletSessionIdByPlayerSessionId(sid);
//        com.cubeia.backoffice.accounting.api.Money walletAmount = convertToWalletMoney(amount);
        try {

        	// Transfer money from static main account to session account.
        	long playerMainAccountId = accountLookupUtil.lookupMainAccountIdForPlayer(new Long(sid.playerId), amount.getCurrencyCode());
        	
        	TransactionBuilder builder = new TransactionBuilder(amount.getCurrencyCode(), amount.getFractionalDigits());
        	builder.entry(playerMainAccountId, amount.getAmount().negate()).
        	entry(walletSessionId, amount.getAmount()).
        	comment("reserve " + amount + " by player " + sid.playerId);
        	
        	TransactionRequest tx = builder.toTransactionRequest();
        	walletService.doTransaction(tx);
        	
//        	log.debug("Sending withdrawal request. " + request);
//            walletService.withdraw(walletAmount, LICENSEE_ID, walletSessionId, "reserve " + amount + " by player " + sid.playerId);

            AccountBalanceResult sessionBalance = walletService.getBalance(walletSessionId);
            Money newBalance = convertFromWalletMoney(sessionBalance.getBalance());

            BalanceUpdate balanceUpdate = new BalanceUpdate(request.getPlayerSessionId(), newBalance, nextId());
            ReserveResponse response = new ReserveResponse(balanceUpdate, amount);
            log.debug("reserve successful: sId = {}, amount = {}, new balance = {}", new Object[]{sid, amount, newBalance});
            response.setProperty(CashGamesBackendService.MARKET_TABLE_SESSION_REFERENCE_KEY, "CUBEIA-MARKET-SID-" + sid.hashCode());
            return response;
        } catch (Exception e) {
            log.error("Failed reserving money", e);
            String msg = "error reserving " + amount + " to session " + walletSessionId + " for player " + sid.playerId + ": " + e.getMessage();

            throw new ReserveFailedException(msg, e, ErrorCode.UNSPECIFIED_FAILURE, true);
        }
    }

    /**
     * Convert from wallet money type to backend money type.
     *
     * @param amount wallet money amount
     * @return converted amount
     */
    private Money convertFromWalletMoney(com.cubeia.backoffice.accounting.api.Money amount) {
        return new Money(amount.getAmount(), new Currency(amount.getCurrencyCode(), amount.getFractionalDigits()));
    }

    /**
     * Convert from backend money type to wallet money type.
     *
     * @param amount amount to convert
     * @return converted amount
     */
    private com.cubeia.backoffice.accounting.api.Money convertToWalletMoney(Money amount) {
        return new com.cubeia.backoffice.accounting.api.Money(amount.getCurrencyCode(), amount.getFractionalDigits(),
                amount.getAmount());
    }

    @Override
    public BatchHandResponse batchHand(BatchHandRequest request) throws BatchHandFailedException {
        try {
            TransactionBuilder txBuilder = createTransactionBuilder(request);

            // Add entries.
            HashMap<Long, PlayerSessionId> sessionToPlayerSessionMap = new HashMap<Long, PlayerSessionId>();
            createHandResultEntries(request, txBuilder, sessionToPlayerSessionMap);
            createRakeEntry(request, txBuilder);

            // Perform transaction.
            TransactionRequest txRequest = txBuilder.toTransactionRequest();
            log.debug("sending tx request to wallet: {}", txRequest);
            TransactionResult txResult = walletService.doTransaction(txRequest);

            // Return result.
            String currencyCode = request.getTotalRake().getCurrencyCode();
            return createBatchHandResponse(sessionToPlayerSessionMap, txResult, currencyCode);
        } catch (UnbalancedTransactionException ute) {
            throw new BatchHandFailedException("error reporting hand result", ute);
        } catch (Exception e) {
            throw new BatchHandFailedException("error reporting hand result", e);
        }
    }

    private TransactionBuilder createTransactionBuilder(BatchHandRequest request) {
        String currencyCode = request.getTotalRake().getCurrencyCode();
        int fractionalDigits = request.getTotalRake().getFractionalDigits();
        return new TransactionBuilder(currencyCode, fractionalDigits);
    }

    private void createRakeEntry(BatchHandRequest request, TransactionBuilder txBuilder) throws BatchHandFailedException {
        //txBuilder.entry(rakeAccountId, convertToWalletMoney(request.getTotalRake()).getAmount());
        TreeMap<Integer, Money> operatorRake = new TreeMap<Integer, Money>();
        for (HandResult hr : request.getHandResults()) {
            int operatorId = hr.getOperator();
            if (operatorRake.containsKey(operatorId)) {
                operatorRake.put(operatorId, operatorRake.get(operatorId).add(hr.getRake()));
            } else {
                operatorRake.put(operatorId, hr.getRake());
            }
        }
        for (Map.Entry<Integer, Money> rakeEntry: operatorRake.entrySet()) {
            String currencyCode = rakeEntry.getValue().getCurrencyCode();
			long rakeAccountId = getRakeAccount(rakeEntry.getKey(), currencyCode);
			log.debug("transferring rake: operatorId = " + rakeEntry.getKey() + ", amount = " + rakeEntry.getValue()+" to RakeAcccountID = "+rakeAccountId);
			txBuilder.entry(rakeAccountId, convertToWalletMoney(rakeEntry.getValue()).getAmount());
			
//            
//            try {
//                // long lookupOperatorAccountId = accountLookupUtil.lookupOperatorAccountId(walletService, rakeEntry.getKey(), rakeEntry.getValue().getCurrencyCode());
//            	long operatorRakeAccountId = accountLookupUtil.lookupOperatorRakeAccountId(rakeEntry.getKey(), rakeEntry.getValue().getCurrencyCode());
//            	log.debug("Transfer rake ["+rakeEntry.getValue()+"] to operator["+rakeEntry.getKey()+"] account["+operatorRakeAccountId+"]");
//				txBuilder.entry(operatorRakeAccountId, convertToWalletMoney(rakeEntry.getValue()).getAmount());
//            } catch (NoSuchAccountException e) {
//            	log.warn("No operator rake account found for rake entry. Will use system rake account as placeholder. Reported error: "+e);
//                txBuilder.entry(getSystemRakeAccount(rakeEntry.getValue().getCurrencyCode()), convertToWalletMoney(rakeEntry.getValue()).getAmount());
//            }
        }
        txBuilder.comment("poker hand result");
        txBuilder.attribute("pokerTableId", String.valueOf((request.getTableId()).integrationId)).attribute("pokerGameId", String.valueOf(GAME_ID)).attribute(
                "pokerHandId", request.getHandId());
    }

    
    /**
     * This method will lookup the correct rake account for the given operator and currency. If a specific operator rake account
     * is found then it will be returned, otherwise a system wide rake account will be looked up. 
     * @param operatorId
     * @param currencyCode
     * @return account id to use
     */
    private long getRakeAccount(Integer operatorId, String currencyCode) {
    	long accountId = -1;
    	try {
    		accountId = accountLookupUtil.lookupOperatorAccount(operatorId, currencyCode, AccountRole.RAKE);
        } catch (NoSuchAccountException e) {
        	log.info("No operator rake account found for rake entry. Will use system rake account as placeholder. Operator["+operatorId+"] Currency["+currencyCode+"]");
        	accountId = getSystemRakeAccount(currencyCode);
        }
    	return accountId;
    }
    
    /**
     * FIXME: This lookup should be by operator and currency and not only currency.
     *       I.e. add support for operator specific rake accounts.
     *       
     * @param currencyCode
     * @return Account id
     */
    private long getSystemRakeAccount(String currencyCode) {
        if (!systemRakeAccounts.containsKey(currencyCode)) {
            long value;
        	value = accountLookupUtil.lookupSystemAccount(currencyCode, AccountRole.RAKE);
            systemRakeAccounts.put(currencyCode, value);
        }
        return systemRakeAccounts.get(currencyCode);
    }

    private long getPromotionsAccount(String currencyCode) {
        if (!promotionsAccounts.containsKey(currencyCode)) {
            long value;
            value = accountLookupUtil.lookupSystemAccount(currencyCode, AccountRole.PROMOTION);
            promotionsAccounts.put(currencyCode, value);
        }
        return promotionsAccounts.get(currencyCode);
    }

    private void createHandResultEntries(BatchHandRequest request, TransactionBuilder txBuilder, HashMap<Long, PlayerSessionId> sessionToPlayerSessionMap) {
        // Add one entry for each hand result.
        for (HandResult hr : request.getHandResults()) {
            log.debug("recording hand result: handId = {}, sessionId = {}, bets = {}, wins = {}, rake = {}",
                    new Object[]{request.getHandId(), hr.getPlayerSession(), hr.getAggregatedBet(), hr.getWin(), hr.getRake()});

            Money resultingAmount = hr.getWin().subtract(hr.getAggregatedBet());

            Long walletSessionId = getWalletSessionIdByPlayerSessionId(hr.getPlayerSession());
            sessionToPlayerSessionMap.put(walletSessionId, hr.getPlayerSession());
            txBuilder.entry(walletSessionId, convertToWalletMoney(resultingAmount).getAmount());
        }
    }

    private BatchHandResponse createBatchHandResponse(HashMap<Long, PlayerSessionId> sessionToPlayerSessionMap, TransactionResult txResult, String currency) {
        List<TransactionUpdate> resultingBalances = new ArrayList<TransactionUpdate>();
        for (AccountBalanceResult sb : txResult.getBalances()) {
            // if (sb.getAccountId() != getSystemRakeAccount(currency) && sessionToPlayerSessionMap.containsKey(sb.getAccountId())) {
        	
        	// If not a player at the table it is probably a rake account and we must skip it
        	if (sessionToPlayerSessionMap.containsKey(sb.getAccountId())) {
                PlayerSessionId playerSessionId = sessionToPlayerSessionMap.get(sb.getAccountId());
                Money balance = convertFromWalletMoney(sb.getBalance());
                BalanceUpdate balanceUpdate = new BalanceUpdate(playerSessionId, balance, nextId());
                resultingBalances.add(new TransactionUpdate(new TransactionId(txResult.getTransactionId()), balanceUpdate));
            }
        }
        return new BatchHandResponse(resultingBalances);
    }

    @Override
    public Money getAccountBalance(int playerId, String currency) throws GetBalanceFailedException {
        long accountId = this.accountLookupUtil.lookupMainAccountIdForPlayer(new Long(playerId), currency);
        log.debug("Found account ID {} for player {}", accountId, playerId);
        if (accountId == -1) {
            log.warn("No account found for " + playerId + " and currency " + currency + ". Returning zero money.");
            return new Money(BigDecimal.ZERO, new Currency(currency, 2));
        }
        Money money = convertFromWalletMoney(walletService.getBalance(accountId).getBalance());
        log.debug("Found balance {} for player {}", money, playerId);
        return money;
    }

    @Override
    public BalanceUpdate getSessionBalance(PlayerSessionId sessionId) throws GetBalanceFailedException {
        AccountBalanceResult sessionBalance = walletService.getBalance(getWalletSessionIdByPlayerSessionId(sessionId));
        Money balanceMoney = convertFromWalletMoney(sessionBalance.getBalance());
        return new BalanceUpdate(sessionId, balanceMoney, nextId());
    }

    @Override
    public void transfer(TransferMoneyRequest request) {
		String currency = request.amount.getCurrencyCode();
		
		// Check if bonus transfer and redefine the request accordingly
		if (request.isToBonusAccount()) {
    		log.debug("Transfer request for bonus account: {}", request);
    		// We need to override the to session account to bonus account . 
    		Long playerId = new Long(request.toSession.playerId);
    		long bonusAccountIdForPlayer = accountLookupUtil.lookupBonusAccountIdForPlayer(playerId, currency);
    		
    		if (bonusAccountIdForPlayer < 0) {
    			log.warn("Missing bonus account for player["+playerId+"] and currency["+currency+"] for payout! Will try to create a new bonus account and make the transfer to the new account. TransferRequest["+request+"]");
    			bonusAccountIdForPlayer = walletService.createBonusAccount(playerId, currency);
    		} 
    		
    		if (bonusAccountIdForPlayer > 0) {
	    		log.debug("Transfer to bonus account["+bonusAccountIdForPlayer+"] for player["+request.toSession.playerId+"]. Override transfer request with new account assignment");
	    		PlayerSessionId toSession = new PlayerSessionId(request.toSession.playerId, bonusAccountIdForPlayer+"");
				request = new TransferMoneyRequest(request.amount, request.fromSession, toSession, request.comment);
    		} else {
    			sendBonusFailedEvent(request);
    			throw new NoSuchAccountException("Could not find bonus account for player["+playerId+"] and currency["+currency+"]. Even tried to create a new bonus account.");
    		}
    	}
    	
        TransactionBuilder txBuilder = new TransactionBuilder(currency, request.amount.getFractionalDigits());
        txBuilder.entry(getWalletSessionIdByPlayerSessionId(request.fromSession), convertToWalletMoney(request.amount.negate()).getAmount());
        txBuilder.entry(getWalletSessionIdByPlayerSessionId(request.toSession), convertToWalletMoney(request.amount).getAmount());
        txBuilder.toTransactionRequest();
        txBuilder.comment(request.comment);
        TransactionRequest txRequest = txBuilder.toTransactionRequest();
        log.debug("sending tx request to wallet: {}", txRequest);
        TransactionResult txResult = walletService.doTransaction(txRequest);
        log.debug("Result: " + txResult);
    }

	private void sendBonusFailedEvent(TransferMoneyRequest request) {
		try {
			SystemEvent event = new SystemEvent();
			event.type = SystemEventType.error.name();
			event.level = SystemLevels.ERROR.name();
			event.name = "Bonus account payout failed";
			event.information = "Failed to lookup and create bonus account. The transfer have not been successfully executed and an error has been propagated (which cause a tournament to stop unexpectedly). Transfer Request: "+request;
			domainEventService.sendEvent(event);
		} catch (Exception e) {
			log.error("Failed to send System Event", e);
		}
	}

    @Override
    public Currency getCurrency(String currencyCode) {
        com.cubeia.backoffice.wallet.api.dto.Currency currency = walletService.getCurrency(currencyCode);
        return new Currency(currency.getCode(),currency.getFractionalDigits());
    }

    @Override
    public void transferMoneyFromPromotionsAccount(PlayerSessionId toAccount, Money money, String comment) {
        TransactionBuilder txBuilder = new TransactionBuilder(money.getCurrencyCode(), money.getFractionalDigits());
        txBuilder.entry(getPromotionsAccount(money.getCurrencyCode()), convertToWalletMoney(money.negate()).getAmount());
        txBuilder.entry(getWalletSessionIdByPlayerSessionId(toAccount), convertToWalletMoney(money).getAmount());
        txBuilder.comment(comment);
        TransactionRequest txRequest = txBuilder.toTransactionRequest();
        log.debug("sending tx request to wallet: {}", txRequest);
        TransactionResult txResult = walletService.doTransaction(txRequest);
        log.debug("Result: " + txResult);
    }

    @Override
    public void transferMoneyToRakeAccount(PlayerSessionId fromAccount, Money money, String comment) {
        TransactionBuilder txBuilder = new TransactionBuilder(money.getCurrencyCode(), money.getFractionalDigits());
        txBuilder.entry(getWalletSessionIdByPlayerSessionId(fromAccount), convertToWalletMoney(money.negate()).getAmount());
        long rakeAccount = getRakeAccount(getOperatorId(fromAccount.playerId), money.getCurrencyCode());
        txBuilder.entry(rakeAccount, convertToWalletMoney(money).getAmount());
        txBuilder.comment(comment);
        TransactionRequest txRequest = txBuilder.toTransactionRequest();
        log.debug("sending tx request to wallet: {}", txRequest);
        TransactionResult txResult = walletService.doTransaction(txRequest);
        log.debug("Result: " + txResult);
    }

    @Override
    public void transferMoneyFromRakeAccount(PlayerSessionId fromAccount, Money money, String comment) {
        TransactionBuilder txBuilder = new TransactionBuilder(money.getCurrencyCode(), money.getFractionalDigits());
        long rakeAccount = getRakeAccount(getOperatorId(fromAccount.playerId), money.getCurrencyCode());
		txBuilder.entry(rakeAccount, convertToWalletMoney(money.negate()).getAmount());
        txBuilder.entry(getWalletSessionIdByPlayerSessionId(fromAccount), convertToWalletMoney(money).getAmount());
        txBuilder.comment(comment);
        TransactionRequest txRequest = txBuilder.toTransactionRequest();
        log.debug("sending tx request to wallet: {}", txRequest);
        TransactionResult txResult = walletService.doTransaction(txRequest);
        log.debug("Result: " + txResult);
    }
    
    private Integer getOperatorId(int playerId) {
        Integer operatorId = clientRegistry.getOperatorId(playerId);
        if(operatorId==null){
            User userById = userService.getUserById(playerId);
            operatorId = userById.getOperatorId().intValue();
        } else {
            log.debug("Operator id not found in client registry for player " + playerId);
        }
        return operatorId;
    }
}
